Refactor legacy protocol helpers into dedicated modules
